refactor: Réorganisation des fichiers et mise à jour des imports dans le module cache  - Renommage et déplacement de plusieurs fichiers pour améliorer la structure du projet.  - Mise à jour des chemins d'importation pour refléter les nouveaux emplacements des fichiers.
